+++ lib/e_way_client.rb
@@ -11,10 +11,13 @@ require "e_way_client/requests/base_request"
 require "e_way_client/requests/query_list_banks_request"
 require "e_way_client/requests/query_list_provinces_request"
 require "e_way_client/requests/send_transaction_request"
+require "e_way_client/requests/query_txn_status_request"
 require "e_way_client/responses/base_response"
 require "e_way_client/responses/query_list_banks_response"
 require "e_way_client/responses/query_list_provinces_response"
 require "e_way_client/responses/send_transaction_response"
+require "e_way_client/responses/query_txn_status_response"
+require "e_way_client/services/gen_action_name"
 
 module EWayClient
 
@@ -25,6 +28,8 @@ module EWayClient
     has :username, classes: String
     has :password, classes: String
     has :secret, classes: String
+    has :log, values: [true, false], default: false
+    has :logger
   end
 
 end