+++ lib/e_way_client/requests/query_list_banks_request.rb
@@ -1,15 +1,13 @@
 module EWayClient
   class QueryListBanksRequest < BaseRequest
 
-    def call
-      SOAPClient.(
-        wsdl: wsdl,
-        action: :query_list_banks,
-        message: {
-          "USERNAME" => username,
-          "PASSWORD" => password,
-        },
-      )
+    private
+
+    def default_message
+      {
+        "USERNAME" => username,
+        "PASSWORD" => password,
+      }
     end
 
   end