+++ lib/e_way_client/requests/query_list_provinces_request.rb
@@ -1,16 +1,5 @@
 module EWayClient
   class QueryListProvincesRequest < BaseRequest
 
-    def call
-      SOAPClient.(
-        wsdl: wsdl,
-        action: :query_list_provinces,
-        message: {
-          "USERNAME" => username,
-          "PASSWORD" => password,
-        },
-      )
-    end
-
   end
 end