+++ lib/e_way_client/services/gen_action_name.rb
@@ -1,0 +1,9 @@
+module EWayClient
+  class GenActionName
+
+    def self.call(klass)
+      klass.name.demodulize.gsub(/Request$/, "").underscore.to_sym
+    end
+
+  end
+end