+++ spec/acceptance/query_txn_status_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe "SendTransaction", vcr: {record: :once} do
+RSpec.describe "QueryTxnStatus", vcr: {record: :once} do
 
   let(:transaction_id) { SecureRandom.hex(6) }
 
@@ -11,7 +11,7 @@ RSpec.describe "SendTransaction", vcr: {record: :once} do
     expect(query_list_banks_response).to be_success
     bank = query_list_banks_response.banks.first
 
-    response = client.send_transaction(
+    send_transaction_response = client.send_transaction(
       transaction_id: transaction_id,
       transaction_date: Time.now,
       transaction_currency: "VND",
@@ -29,10 +29,14 @@ RSpec.describe "SendTransaction", vcr: {record: :once} do
       bank_id: bank.bank_id,
       bank_account_number: "0123456789",
       bankbranch_name: "Test Branch",
-      bankbranch_address: "123 Testy"
+      bankbranch_address: "Test address"
     )
 
+    expect(send_transaction_response).to be_success
+
+    response = client.query_txn_status(transaction_id)
     expect(response).to be_success
+    expect(%w(P S C W).include?(response.transaction_status)).to be true
   end
 
 end