+++ spec/lib/e_way_client/services/gen_action_name_spec.rb
@@ -1,0 +1,17 @@
+require 'spec_helper'
+
+module EWayClient
+  RSpec.describe GenActionName do
+
+    describe ".call" do
+      it "generates the action name based on the request class" do
+        expect(described_class.(SendTransactionRequest)).
+          to eq :send_transaction
+
+        expect(described_class.(QueryListBanksRequest)).
+          to eq :query_list_banks
+      end
+    end
+
+  end
+end