+++ spec/spec_helper.rb
@@ -5,10 +5,14 @@ require "pathname"
 require "active_support/core_ext/hash/indifferent_access"
 require "yaml"
 require "virtus-matchers"
+require "timecop"
 
 SPEC_DIR = Pathname.new(File.dirname(__FILE__))
 CONFIG_YML = SPEC_DIR.join("config.yml")
-CONFIG = YAML.load_file(CONFIG_YML).with_indifferent_access
+CONFIG = YAML.load_file(CONFIG_YML).with_indifferent_access.merge(
+  log: false,
+  logger: Logger.new(STDOUT)
+)
 
 require 'vcr'