+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [TBA]
+### Added
+- Option to log Savon requests
+- `query_txn_status` call to get a transaction's status
+
 ## [0.1.1] - 2017-01-09
 ### Fixed
 - Do not blow up if `diffgram` in XML response is missing. Return nil instead.